requirements.txt
Flask==3.0.3
Pillow==10.4.0
torch==2.3.1
torch==2.4.0
torchvision==0.19.0
transformers==4.44.0
